+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## 2020-06-03
+
+## Added
+
+- Added Oxford covid-19 government response tracker pipeline 
+
 ## 2020-06-01
 
 ## Added